nightly

+++ application/models/campaign_m.php
@@ -24,11 +24,13 @@ Class Campaign_m extends CI_Model {
 	  if($result = $this->db->get_row($sql,'ARRAY_A')){
 	    $result = array_merge($result,$this->getStatus($result));
 		//Merge Assets if exists
-		if($assets = getAssets($result['GID'])){
+		if($assets = $this->getAssets($result['GID'])){
 		  $result = array_merge($result,$assets);
 		}
 		//Merge extra Page if exists
-		//TO DO 
+		if($pages = $this->getPages($result['GID'])){
+		  $result = array_merge($result,$pages);
+		}
 	  }else{
 		$result = null;
 	  }	  
@@ -97,7 +99,14 @@ Class Campaign_m extends CI_Model {
   }
   
   public function getPages($GID){
-  
+  $this->load->model('page_m','page');
+    $data = array();
+	if($pages = $this->page->retrievePage(array('campaign_page.GID'=>$GID))){
+		foreach ($pages as $page){
+			$data['pages'][] = array('id'=>$page['page_id'],'name'=>$page['page_short_name'],'url'=>menu_url('page/'.$page['page_id'])); 
+		}
+	}
+	return $data;
   }
   
   public function setStatus($on_wait = false,$on_progress = false,$on_upload = false,$on_vote = false,$on_judging = false,$is_off = true){
@@ -161,7 +170,17 @@ Class Campaign_m extends CI_Model {
   
   public function removeCampaign($gid)
   {
+    
+	 $this->load->model('media_m','media');
+	 //Archive Campaign if it's already a participant
+	 if($media = $this->media->mediaByRandom($gid)){
+		return false;
+	 }
+   
 	$deleted = $this->db->query("DELETE FROM campaign_group WHERE GID = ".$gid);
+	$deleted = $this->db->query("DELETE FROM campaign_page WHERE GID = ".$gid);
+	$deleted = $this->db->query("DELETE FROM campaign_group_assets WHERE GID = ".$gid);
+	
 	if($deleted){
 		if(is_dir(CUSTOMER_IMAGE_DIR.$gid)){
 			if(!rm_all_dir(CUSTOMER_IMAGE_DIR.$gid)){
@@ -228,7 +247,20 @@ Class Campaign_m extends CI_Model {
    $sql  = "SELECT campaign_group.* ";
    $sql .= "FROM campaign_group ";
    $sql .= "WHERE campaign_group.GID = ".$gid;
-   return $this->db->get_row($sql,'ARRAY_A');
+   $result = null;
+   	  if($result = $this->db->get_row($sql,'ARRAY_A')){
+	    $result = array_merge($result,$this->getStatus($result));
+		//Merge Assets if exists
+		if($assets = $this->getAssets($result['GID'])){
+		  $result = array_merge($result,$assets);
+		}
+		//Merge extra Page if exists
+		if($pages = $this->getPages($result['GID'])){
+		  $result = array_merge($result,$pages);
+		}
+	  }	  
+   
+    return $result;
   }
   
   public function setStatusCampaign($gid,$status)
